Seperated PROVIDER to lean code in App.js
